Fix --doc option of LilyPond UNIX installers
As reported at
https://lists.gnu.org/archive/html/lilypond-user/2020-07/msg00119.html
and tracked at
https://gitlab.com/lilypond/lilypond/-/issues/3768,
when the installers are called with qualified path, the --doc option
stops working. David Sumbler diagnosed the problem as an incorrect path
for searching the docs on the website. Fix this by extracting the base
name of $me.

A better fix would be to record the version in the script upon
generation to make it work even when renamed. For now, this seems
good enough.